Register default mvc routes first.
